# [db] Optimized request polling cost