<!-- Describe the changes in this PR -->

close https://github.com/skypilot-org/skypilot/issues/6945

According profile, there is a non-trivial CPU cost of `get_request_async` call since it is called in busy loops (`/api/get`, `/api/stream`), this PR simplifies the SQL to check status.

Test:

```bash
python test.py
10000 times get_request_status Time taken: 3.0183849334716797
10000 times get_request_status_async Time taken: 0.4276120662689209
```

About 7x improvement.

```python
$ cat test.py
import asyncio
import time
from sky.server.requests import requests

async def main():
    now = time.time()
    for _ in range(10000):
        req_status = await requests.get_request_async('skypilot-status-refresh-daemon')
    end = time.time()
    print(f'10000 times get_request_async Time taken: {end - now}')

    now = time.time()
    for _ in range(10000):
        req_status = await requests.get_request_status_async('skypilot-status-refresh-daemon')
    end = time.time()
    print(f'10000 times get_request_status_async Time taken: {end - now}')


if __name__ == '__main__':
    asyncio.run(main())
```

Also, `sky status` does not have regression:

```bash
# PR branch
$ multitime sky status
===> multitime results
1: sky status
            Mean        Std.Dev.    Min         Median      Max
real        1.793       0.524       1.394       1.559       2.827
user        1.373       0.316       1.012       1.309       1.952
sys         0.848       0.247       0.527       1.005       1.085
# master branch
$ multitime sky status
===> multitime results
1: sky status
            Mean        Std.Dev.    Min         Median      Max
real        1.800       0.680       1.401       1.496       3.158
user        1.168       0.176       0.952       1.232       1.430
sys         0.911       0.283       0.594       0.751       1.356
```

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: install pre-commit (auto-check on commit) or `bash format.sh`
- [x] Any manual or new tests for this PR (please specify below)
- [ ] All smoke tests: `/smoke-test` (CI) or `pytest tests/test_smoke.py` (local)
- [ ] Relevant individual tests: `/smoke-test -k test_name` (CI) or `pytest tests/test_smoke.py::test_name` (local)
- [ ] Backward compatibility: `/quicktest-core` (CI) or `pytest tests/smoke_tests/test_backward_compat.py` (local)

<!-- CI commands (/-prefixed) can only be triggered by repo members -->